Make strict IntMap merges strict
* Make `Data.IntMap.Merge.Strict` tactics (except `preserveMissing`)
  strict.

* Add a strict `Data.Map.Merge.Strict` `preserveMissing'` tactic.
  We may want to just call this `preserveMissing`....

Fixes #609
